fix(quirc): handle QR code errors better
node-quirc will actually resolve the promise even if it finds a QR code it can't decode. This makes sense since there may be QR codes it could decode and you need a way to communicate that. The ones that error will have just an `err` property and nothing else, so we now just return the first QR code that does not have an `err` property.