fix: improve signal to noise ratio of target mark scores
Here's the process:
- “outline” the target ovals to expand them by 1px, i.e. color all pixels black that are adjacent to black pixels
- get the inner rect encompassing the entire inner oval
- figure out how many white pixels could be filled in in that image
- count the number of newly-black pixels in the ballot image compared to the template
- use the ratio of black to white as the score

Additionally, we pick the lowest score after shifting the image 1px in each-direction. All together this means that target ovals that were not marked by a voter have a score of 0 (though in practice I suspect it’ll not always be zero, but close), and anything marked by the voter where they’re clearly trying to do something will have a non-trivial score.