fix: identify targets by width
Previously, only aspect ratio was used to identify targets. This meant that a contest box that had the same aspect ratio as a target would itself be counted as a target. This ensures that only shapes that are the expected percentage width of a contest box are identified as targets.