fix: be smarter about column sizes
We should base the expected contest width on the number of columns by default. When there are three columns, we now expect a contest box to be between 28-39% vs 20-40% previously. If only two columns are given, it'd be 45-55%. More work will likely need to be done to support two-column layouts, but this is a fairly obvious start.