fix: handle really skewed ballots better
When looking for contests we look for both the bounding box and the actual corners. If the rectangle were perfectly straight, the corners would be the corners of the bounding box (and that's exactly what we see with the template). But for really skewed images, they can be pretty different points. My prior testing on using the corners as input to the homography was not done on sufficiently skewed ballots, so I was not seeing much benefit. I believe the issue we're seeing now is due to the extremely long contest, which allows the skew to create a rather large difference in the bounding box corners from the real corners. By using the real detected corners as the input to the homography, we're able to get much better results.

However, there was also a bug in the corner finding code. This fixes the algorithm to stop early if the resulting corner would imply more skew than we're willing to accept. By default we allow up to 5° rotation skew.