fix: prevent infinite loop looking for corners
I wasn't doing proper bounds checking on the x,y values, so in some conditions we might go beyond the bounding box looking for black pixels. In some cases, we'd even go outside the bounds of the image. Once that happens, we'd loop forever since the pixel value at such points is `undefined`, which will never equal 0.

This happened because we were looking for TWO consecutive black pixels along the edge of the bounding box in either direction. However, there only has to be ONE consecutive pixel for it to push the bounding box out. This is the case that would cause the corner finder to keep going beyond the bounding box and possibly forever.

Finding two consecutive pixels was from before I implemented backtracking, and no longer seems necessary. Once we find a single black pixel we can backtrack to find the real corner. Additionally, I added bounds checking since we have no reason to leave the bounding box looking for corners.