fix: correct shape sizes
Off-by-1 errors! This has been wrong for some time, and will likely have minimal effect on scores. Still, it's important to get every little bit we can to aid alignment. Without this, the right and bottom edges of all shapes (i.e. targets and contests) are cut off by one pixel.

Fixing this also exposed another bug. We were continuing to look for contests in a column past the point we expected to see the first one as long as we saw _any_ shape by the point we expected to see a contest box. This is now fixed too.