fix: use all four contest corners on 2-column pages
On 2-column pages the contests are wide enough that the skew on the right side can cause problems with the left side lining up, even though we got the left corners lined up. This has led to spurious overvotes on the tall+wide either-neither contest in the upcoming general election in MS. This change was tested with 3798 scans, and it went from 91 either-neither overvotes to 2 either-neither overvotes.